refactor:[#259] session/config, 순수 도메인 모델 루트로 이동
- session.domain.* (5개) → domain.*
- session.config.* (2개) → config.*
- session.persistence.entity.* (8개) → domain.entity.*
- 전체 패키지 내 관련 import 경로 일괄 수정
