+++ src/main/java/com/ktb/interview/application/service/InterviewFeedbackOrchestratorImpl.java
@@ -16,8 +16,8 @@ import com.ktb.answer.repository.AnswerRepository;
 import com.ktb.interview.port.out.AiInterviewPort;
 import com.ktb.answer.service.AnswerDomainService;
 import com.ktb.interview.application.InterviewFeedbackOrchestrator;
-import com.ktb.interview.session.domain.InterviewHistoryItem;
-import com.ktb.interview.session.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewHistoryItem;
+import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.session.service.InterviewSessionService;
 import com.ktb.common.domain.ErrorCode;
 import com.ktb.hashtag.domain.QuestionHashtag;
