+++ src/main/java/com/ktb/interview/application/service/InterviewFollowUpOrchestratorImpl.java
@@ -12,8 +12,8 @@ import com.ktb.interview.dto.ai.InterviewFollowUpQuestionRequest;
 import com.ktb.interview.dto.ai.InterviewHistoryRequest;
 import com.ktb.interview.port.out.AiInterviewPort;
 import com.ktb.interview.application.InterviewFollowUpOrchestrator;
-import com.ktb.interview.session.domain.InterviewHistoryItem;
-import com.ktb.interview.session.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewHistoryItem;
+import com.ktb.interview.domain.InterviewSession;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
