+++ src/main/java/com/ktb/interview/application/service/flow/FinalFeedbackCompletionService.java
@@ -2,7 +2,7 @@ package com.ktb.interview.application.service.flow;
 
 import com.ktb.interview.dto.ai.InterviewFeedbackDataResponse;
 import com.ktb.interview.port.out.InterviewSessionFinalFeedbackStore;
-import com.ktb.interview.session.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.session.service.InterviewSessionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
