+++ src/main/java/com/ktb/interview/application/service/flow/InterviewRealSubmissionFlowService.java
@@ -6,10 +6,10 @@ import com.ktb.interview.application.InterviewFollowUpOrchestrator;
 import com.ktb.interview.assembler.InterviewRealHistoryAssembler;
 import com.ktb.interview.mapper.InterviewSubmissionResponseMapper;
 import com.ktb.interview.dto.ai.InterviewHistoryRequest;
-import com.ktb.interview.session.config.InterviewSessionProperties;
-import com.ktb.interview.session.domain.InterviewQuestionSnapshot;
-import com.ktb.interview.session.domain.InterviewSession;
-import com.ktb.interview.session.domain.InterviewSessionStatus;
+import com.ktb.interview.config.InterviewSessionProperties;
+import com.ktb.interview.domain.InterviewQuestionSnapshot;
+import com.ktb.interview.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewSessionStatus;
 import com.ktb.interview.session.dto.request.RealAnswerSubmitRequest;
 import com.ktb.interview.session.dto.response.InterviewRealSubmitResponse;
 import com.ktb.interview.session.service.InterviewSessionService;
