+++ src/main/java/com/ktb/interview/application/service/flow/InterviewSessionFeedbackQueryFlowService.java
@@ -7,9 +7,9 @@ import com.ktb.file.dto.request.PresignedUrlMethod;
 import com.ktb.file.dto.request.PresignedUrlRequest;
 import com.ktb.file.dto.response.PresignedUrlResponse;
 import com.ktb.file.service.S3PresignedUrlService;
-import com.ktb.interview.session.domain.InterviewHistoryItem;
-import com.ktb.interview.session.domain.InterviewSession;
-import com.ktb.interview.session.domain.InterviewSessionStatus;
+import com.ktb.interview.domain.InterviewHistoryItem;
+import com.ktb.interview.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewSessionStatus;
 import com.ktb.interview.session.dto.response.InterviewFinalFeedbackMetricResponse;
 import com.ktb.interview.session.dto.response.InterviewHistoryResponse;
 import com.ktb.interview.session.dto.response.InterviewSessionBadCaseFeedbackResponse;
