+++ src/main/java/com/ktb/interview/application/service/flow/InterviewSessionFinalFeedbackFlowService.java
@@ -6,10 +6,10 @@ import com.ktb.hashtag.repository.QuestionHashtagRepository;
 import com.ktb.interview.application.InterviewFeedbackOrchestrator;
 import com.ktb.interview.dto.ai.InterviewFeedbackDataResponse;
 import com.ktb.interview.mapper.InterviewSubmissionResponseMapper;
-import com.ktb.interview.session.domain.InterviewHistoryItem;
-import com.ktb.interview.session.domain.InterviewSession;
-import com.ktb.interview.session.domain.InterviewSessionFeedback;
-import com.ktb.interview.session.domain.InterviewSessionStatus;
+import com.ktb.interview.domain.InterviewHistoryItem;
+import com.ktb.interview.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewSessionFeedback;
+import com.ktb.interview.domain.InterviewSessionStatus;
 import com.ktb.interview.session.dto.response.InterviewSessionFinalFeedbackResponse;
 import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
 import com.ktb.interview.session.mapper.InterviewSessionFeedbackMapper;
