+++ src/main/java/com/ktb/interview/domain/entity/InterviewTurnId.java
@@ -1,4 +1,4 @@
-package com.ktb.interview.session.persistence.entity;
+package com.ktb.interview.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -10,18 +10,18 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 토픽 피드백 복합키(session_id + topic_id).
+ * 인터뷰 턴 복합키(session_id + turn_order).
  */
 @Embeddable
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EqualsAndHashCode
-public class InterviewTopicFeedbackId implements Serializable {
+public class InterviewTurnId implements Serializable {
 
     @Column(name = "session_id", nullable = false, length = 64)
     private String sessionId;
 
-    @Column(name = "topic_id", nullable = false)
-    private Integer topicId;
+    @Column(name = "turn_order", nullable = false)
+    private Integer turnOrder;
 }
