+++ src/main/java/com/ktb/interview/mapper/InterviewSubmissionResponseMapper.java
@@ -7,9 +7,9 @@ import com.ktb.interview.dto.ai.InterviewFeedbackDataResponse;
 import com.ktb.interview.dto.ai.InterviewKeywordResultResponse;
 import com.ktb.interview.dto.ai.InterviewOverallFeedbackResponse;
 import com.ktb.interview.dto.ai.InterviewTopicFeedbackResponse;
-import com.ktb.interview.session.domain.InterviewHistoryItem;
-import com.ktb.interview.session.domain.InterviewQuestionSnapshot;
-import com.ktb.interview.session.domain.InterviewSessionStatus;
+import com.ktb.interview.domain.InterviewHistoryItem;
+import com.ktb.interview.domain.InterviewQuestionSnapshot;
+import com.ktb.interview.domain.InterviewSessionStatus;
 import com.ktb.interview.session.dto.response.InterviewFinalFeedbackMetricResponse;
 import com.ktb.interview.session.dto.response.InterviewHistoryResponse;
 import com.ktb.interview.session.dto.response.InterviewNextQuestionTurnResponse;
