+++ src/main/java/com/ktb/interview/session/mapper/InterviewSessionFeedbackMapper.java
@@ -6,7 +6,7 @@ import com.ktb.interview.dto.ai.InterviewFeedbackMetricResponse;
 import com.ktb.interview.dto.ai.InterviewKeywordResultResponse;
 import com.ktb.interview.dto.ai.InterviewOverallFeedbackResponse;
 import com.ktb.interview.dto.ai.InterviewTopicFeedbackResponse;
-import com.ktb.interview.session.domain.InterviewSessionFeedback;
+import com.ktb.interview.domain.InterviewSessionFeedback;
 import java.util.List;
 
 /**
