+++ src/main/java/com/ktb/interview/session/persistence/adapter/JpaInterviewSessionFinalFeedbackStore.java
@@ -8,15 +8,15 @@ import com.ktb.interview.dto.ai.InterviewKeywordResultResponse;
 import com.ktb.interview.dto.ai.InterviewOverallFeedbackResponse;
 import com.ktb.interview.dto.ai.InterviewTopicFeedbackResponse;
 import com.ktb.interview.port.out.InterviewSessionFinalFeedbackStore;
-import com.ktb.interview.session.domain.InterviewHistoryItem;
-import com.ktb.interview.session.domain.InterviewQuestionSnapshot;
-import com.ktb.interview.session.domain.InterviewSession;
+import com.ktb.interview.domain.InterviewHistoryItem;
+import com.ktb.interview.domain.InterviewQuestionSnapshot;
+import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
-import com.ktb.interview.session.persistence.entity.InterviewSessionEntity;
-import com.ktb.interview.session.persistence.entity.InterviewSessionFeedbackEntity;
-import com.ktb.interview.session.persistence.entity.InterviewSessionMetricEntity;
-import com.ktb.interview.session.persistence.entity.InterviewTopicFeedbackEntity;
-import com.ktb.interview.session.persistence.entity.InterviewTurnEntity;
+import com.ktb.interview.domain.entity.InterviewSessionEntity;
+import com.ktb.interview.domain.entity.InterviewSessionFeedbackEntity;
+import com.ktb.interview.domain.entity.InterviewSessionMetricEntity;
+import com.ktb.interview.domain.entity.InterviewTopicFeedbackEntity;
+import com.ktb.interview.domain.entity.InterviewTurnEntity;
 import com.ktb.interview.session.persistence.repository.InterviewSessionEntityRepository;
 import com.ktb.interview.session.persistence.repository.InterviewSessionFeedbackEntityRepository;
 import com.ktb.interview.session.persistence.repository.InterviewSessionMetricEntityRepository;
