+++ src/main/java/com/ktb/interview/session/persistence/repository/InterviewSessionEntityRepository.java
@@ -1,6 +1,6 @@
 package com.ktb.interview.session.persistence.repository;
 
-import com.ktb.interview.session.persistence.entity.InterviewSessionEntity;
+import com.ktb.interview.domain.entity.InterviewSessionEntity;
 import com.ktb.interview.session.persistence.repository.projection.InterviewSessionFinalFeedbackReadModel;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
