+++ src/main/java/com/ktb/interview/session/persistence/repository/InterviewSessionFeedbackEntityRepository.java
@@ -1,6 +1,6 @@
 package com.ktb.interview.session.persistence.repository;
 
-import com.ktb.interview.session.persistence.entity.InterviewSessionFeedbackEntity;
+import com.ktb.interview.domain.entity.InterviewSessionFeedbackEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InterviewSessionFeedbackEntityRepository extends JpaRepository<InterviewSessionFeedbackEntity, String> {
