+++ src/main/java/com/ktb/interview/session/persistence/repository/InterviewSessionMetricEntityRepository.java
@@ -1,7 +1,7 @@
 package com.ktb.interview.session.persistence.repository;
 
-import com.ktb.interview.session.persistence.entity.InterviewSessionMetricEntity;
-import com.ktb.interview.session.persistence.entity.InterviewSessionMetricId;
+import com.ktb.interview.domain.entity.InterviewSessionMetricEntity;
+import com.ktb.interview.domain.entity.InterviewSessionMetricId;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
