+++ src/main/java/com/ktb/interview/session/persistence/repository/InterviewTurnEntityRepository.java
@@ -1,7 +1,7 @@
 package com.ktb.interview.session.persistence.repository;
 
-import com.ktb.interview.session.persistence.entity.InterviewTurnEntity;
-import com.ktb.interview.session.persistence.entity.InterviewTurnId;
+import com.ktb.interview.domain.entity.InterviewTurnEntity;
+import com.ktb.interview.domain.entity.InterviewTurnId;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
