+++ src/main/java/com/ktb/interview/session/repository/impl/RedisInterviewSessionFeedbackRepository.java
@@ -1,7 +1,7 @@
 package com.ktb.interview.session.repository.impl;
 
-import com.ktb.interview.session.config.InterviewSessionStorePolicy;
-import com.ktb.interview.session.domain.InterviewSessionFeedback;
+import com.ktb.interview.config.InterviewSessionStorePolicy;
+import com.ktb.interview.domain.InterviewSessionFeedback;
 import com.ktb.interview.session.repository.InterviewSessionFeedbackRepository;
 import java.time.LocalDateTime;
 import java.util.Map;
