+++ src/main/java/com/ktb/interview/session/repository/impl/RedisInterviewSessionRepository.java
@@ -1,7 +1,7 @@
 package com.ktb.interview.session.repository.impl;
 
-import com.ktb.interview.session.config.InterviewSessionStorePolicy;
-import com.ktb.interview.session.domain.InterviewSession;
+import com.ktb.interview.config.InterviewSessionStorePolicy;
+import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.session.repository.InterviewSessionRepository;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
